Code duplication removal. I should stop creating alias methods that I am going to end up removing.
What I think is a fix for the signs not updating properly. We'll find out in the long run.

Made it so that wormhole owner updates the gate name sign.

Updated the way we handled lever state changes to properly use bitwise math. None of my funky byte to int to byte bs. :)

Now when a user cancels a dial during light up animation it properly darkens the gate and doesn't lock the gate in a zombie state.